+++ controllers/DisqusController.php
@@ -3,13 +3,17 @@ namespace Craft;
 
 class DisqusController extends BaseController
 {
-	protected $allowAnonymous = array('logoutRedirect');
+    /**
+     * @var array
+     */
+    protected $allowAnonymous = array('logoutRedirect');
 
+    /**
+     * Log the current user out
+     */
     public function actionLogoutRedirect()
     {
         craft()->userSession->logout(false);
         $this->redirect($_SERVER['HTTP_REFERER']);
     }
-
 }
-?>